feat(cli): add cli.py file
add cli.py file
