+++ docs/aktools_version.md
@@ -12,6 +12,8 @@
 
 ## 开发进度
 
+0.0.77: add: add cli.py file
+
 0.0.76: fix: fix rename master to main
 
 0.0.75: add: add version interface
